feat(components): make ranking prompt a component
