+++ src/components/RankingPrompt/RankingPrompt.module.css
@@ -1,10 +1,3 @@
-.wrapper {
-  min-height: 100vh;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
-
 .mainBox {
   color: var(--common-white);
   background-color: var(--primary-dark);
@@ -36,17 +29,4 @@
   color: var(--secondary-dark);
   font-size: var(--question-text);
   font-weight: 600;
-}
-
-.logoWrapper {
-  width: 100%;
-  position: absolute;
-  bottom: 0;
-  left: 0;
-  transform: translateY(65%);
-}
-
-.logo {
-  width: 65%;
-  margin: 0 auto;
 }
