+++ src/components/index.js
@@ -7,3 +7,4 @@ export * from './BackgroundDecoration';
 export * from './QuestionCounter';
 export * from './Input';
 export * from './PopupClose';
+export * from './RankingPrompt';
