+++ src/pages/Answers/Answers.module.css
@@ -5,7 +5,11 @@
 
 .popupOverlay {
   grid-area: 1 / -1 / 1 / -1;
-  /* background-color: rgba(0, 0, 0, 0.5); */
+  background-color: rgba(0, 0, 0, 0.75);
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  z-index: 1;
 }
 
 .content {
