+++ src/pages/index.js
@@ -1,14 +1,12 @@
 import { Home } from './Home';
 import { Loading } from './Loading';
 import { Quiz } from './Quiz';
-import { RankingPrompt } from './RankingPrompt';
 import { Answers } from './Answers';
 
 const pages = Object.freeze({
   home: Home,
   loading: Loading,
   quiz: Quiz,
-  rankingPrompt: RankingPrompt,
   answers: Answers,
 });
